% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvTune_oop.R
\name{coef.cvTune}
\alias{coef.cvTune}
\alias{predict.cvTune}
\title{Coeffcients and predictions by the tuned model in \code{cvTune} objects.}
\usage{
\method{coef}{cvTune}(object, ...)

\method{predict}{cvTune}(
  object,
  newx,
  type = c("link", "response", "coefficients", "nonzero", "class"),
  ...
)
}
\arguments{
\item{object}{an instance of \code{\link{cvTune}}.}

\item{...}{additional arguments passed to \code{\link[glmnet]{coef.glmnet}}
or \code{\link[glmnet]{predict.glmnet}}.}

\item{newx}{matrix of explanatory variables (a numeric matrix or an object
generated by \code{\link[stats]{model.matrix}}).}

\item{type}{of the prediction, as specified by
\code{\link[glmnet]{predict.glmnet}}.}
}
\value{
\code{coef()} returns a sparse one-column matrix with the model
coefficients as specified by \code{\link[glmnet]{coef.glmnet}}.
\code{predict()} returns linear predictor scores, predictions of the response or
class assignment as specified by \code{\link[glmnet]{predict.glmnet}}.
}
\description{
Methods \code{coef()} and \code{predict()} return coefficients and predictions of the
best performing \code{GLMNET} model.
}
