% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modData_oop.R
\name{components.modData}
\alias{components.modData}
\alias{components.modMData}
\title{Access components of a \code{modData} object.}
\usage{
\method{components}{modData}(object, type = c("train", "test", "limits", "stats"), ...)

\method{components}{modMData}(object, type = c("train", "test", "limits", "stats"), ...)
}
\arguments{
\item{object}{a \code{modData} object.}

\item{type}{the requested component of the object, see \code{Details}.}

\item{...}{extra arguments, currently none.}
}
\value{
see: \code{Details}.
}
\description{
Accesses components of the \code{modData} object.
}
\details{
The following are returned for the argument \code{type} specified as:
\itemize{
\item \code{train}: the training numeric matrix with the selected modeling variables
and adjusted for the batch effects by ComBat.
\item \code{test}: the test numeric matrix with the selected modeling variables
and adjusted for the batch effects by ComBat.
\item \code{limits}: cutoffs of median, mean and Gini index used for selection of the
modeling variables.
\item \code{stats}: feature distribution statistics (median, mean, Gini index) prior
to the batch selection (all variables) and following the batch correction via
ComBat (modeling variables only).
}
}
