% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTrain_oop.R
\name{components.modTrain}
\alias{components.modTrain}
\title{Access components of a \code{modTrain} object.}
\usage{
\method{components}{modTrain}(object, type = c("models", "stats", "globals", "errors"), ...)
}
\arguments{
\item{object}{a \code{modTrain} object.}

\item{type}{the requested component of the object, see \code{Details}.}

\item{...}{extra arguments, currently none.}
}
\value{
see: \code{Details}.
}
\description{
Accesses components of the \code{modTrain} object.
}
\details{
The following are returned for the argument \code{type} specified as:
\itemize{
\item \code{models}: \code{cv.glmnet} models.
\item \code{stats}: model performance statistics.
\item \code{globals}: global parameter values used for construction of \code{glmnet} models.
\item \code{errors}: error messages for failed modeling tasks.
}
}
