% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{cvTune}
\alias{cvTune}
\title{\code{cvTune} class.}
\usage{
cvTune(x, ...)
}
\arguments{
\item{x}{a named list with elements described in \code{Details}.}

\item{...}{extra arguments, currently none.}
}
\value{
an instance of the \code{cvTune} class with components
described in \code{Details}.
}
\description{
Creates an object of the \code{modTune} class, which stores results of tuning of
a single \code{GLMNET} model by repeated cross-validation. The instances can be
created by \code{\link{tune_glmnet}} function.
}
\details{
The object has the following components:
\itemize{
\item \code{stats}: model performance statistic, i.e. the cost function, for the
locally optimal \code{lambda} values in repeats of the cross-validation. Those
are the \code{lambda} values, one per repeat, which correspond to the optimum of
the cost function in out-of-fold predictions
\item \code{lambda}: the globally optimal value of \code{lambda}m, i.e. the optimum of
\code{lambdas} for particular repeats listed in \code{stats}
\item \code{model}: a \code{GLMNET} model fit by \code{\link[glmnet]{glmnet}} for the
globally optimal \code{lambda} value
}

Note the methods \code{\link{nobs.cvTune}}, \code{\link{summary.cvTune}},
\code{\link{plot.cvTune}}, \code{\link{coef.cvTune}}, and
\code{\link{predict.cvTune}} defined for the class.
}
