% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_regression}
\alias{evaluate_regression}
\alias{evaluate_classification}
\title{Evaluation of performance of \code{cv.glmnet} models in the training data set.}
\usage{
evaluate_regression(model_lst, lambda, y_lst, x)

evaluate_classification(model_lst, lambda, y_lst, x)
}
\arguments{
\item{model_lst}{a list of \code{cv.glmnet} models.}

\item{lambda}{a vector with lambda values, for which the predictions will
be made.}

\item{y_lst}{a list of named vectors with the observed responses.}

\item{x}{a numeric matrix of explanatory factors.}
}
\value{
\code{evaluate_regression()} returns a list of two numeric vectors, with
Pearson's and Spearman's correlation coefficients, respectively.
\code{evaluate_classification()} returns a list of two numeric vectors ,
with overall accuracy and Cohen's \eqn{\kappa}, respectively.
}
\description{
For regression models, i.e. Gaussian and Poisson family, the function
computes correlation coefficients between the predicted and observed outcome.
For classification models overall accuracy and Cohen's \eqn{\kappa} are
computed for comparison of the actual and predicted class assignment.
}
\details{
Overall accuracy and Cohen's \eqn{\kappa} are computed with
\code{\link[caret]{multiClassSummary}}.
}
