% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_regression}
\alias{fit_regression}
\alias{fit_classification}
\title{Serial fit of models.}
\usage{
fit_regression(y_lst, x, safe = FALSE, ...)

fit_classification(y_lst, x, safe = FALSE, ...)
}
\arguments{
\item{y_lst}{a list of named vectors with the responses.}

\item{x}{a matrix with explanatory factors.}

\item{safe}{logical, if \code{TRUE}, the procedures is run with purrr's
\code{safely()}: correct evaluations are stored in \code{cv.models},
errors are stored in \code{errors}.}

\item{...}{other arguments passed to \code{\link[glmnet]{cv.glmnet}} and
\code{\link[glmnet]{glmnet}}, such as \code{alpha}, \code{family} or \code{type.measure}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{cv_models}: a list \code{cv.glmnet} class models.
\item \code{errors}: a list of error messages.
}
}
\description{
Returns a list of \code{cv.glmnet} class models given a list of responses and
a common matrix of explanatory variables. \code{fit_regression()} is intended for
tuning and training regression models, \code{fit_classification()} is designed
for classification models (binomial and multinomial).
}
\details{
Intended for internal use. Will run in parallel, when a \code{furrr}-compatible
backend is declared.
Classification models are handled in a special way involving call evaluation,
because direct passing of additional arguments
to \code{\link[glmnet]{cv.glmnet}} and \code{\link[glmnet]{glmnet}} via
dots causes C++ errors. This is likely attributed to an unexpected behavior
of the \code{glmnet} \code{lognet()} function.
}
