% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdsc_expression}
\alias{gdsc_expression}
\title{Baseline gene expression of cancer cell lines, GDSC2 study.}
\format{
a numeric matrix of RMA normalized and log-transformed gene
expression levels, genes in rows and cell line identifiers in columns.
The matrix has 17419 rows and 200 columns.
}
\source{
https://osf.io/c6tfx/
}
\usage{
data(gdsc_expression)
}
\description{
RMA-normalized and log-transformed gene expression values in 200 randomly
selected cancer cell lines from the GDSC2 drug treatment screening without
any treatment. The data set may be used as an example training set for
exploring tools of the package. The full data set is available
in the repository by Maeser and colleagues at https://osf.io/c6tfx/.
}
\references{
Yang W, Soares J, Greninger P, Edelman EJ, Lightfoot H, Forbes S,
Bindal N, Beare D, Smith JA, Thompson IR, et al. Genomics of Drug
Sensitivity in Cancer (GDSC): a resource for therapeutic biomarker
discovery in cancer cells. Nucleic Acids Res (2013) 41:D955–D961.
doi:10.1093/NAR/GKS1111

Maeser D, Gruener RF, Huang RS. oncoPredict: an R package for predicting
in vivo or cancer patient drug response and biomarkers from cell line
screening data. Brief Bioinform (2021) 22:1–7. doi:10.1093/BIB/BBAB260

Geeleher P, Cox NJ, Huang RS. Clinical drug response can be predicted
using baseline gene expression levels and in vitro drug sensitivity in
cell lines. Genome Biol (2014) 15: doi:10.1186/gb-2014-15-3-r47

Geeleher P, Cox N, Stephanie Huang R. pRRophetic: An R Package for Prediction
of Clinical Chemotherapeutic Response from Tumor Gene Expression Levels.
PLoS One (2014) 9:e107468. doi:10.1371/JOURNAL.PONE.0107468
}
