% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvTune_oop.R
\name{is_cvTune}
\alias{is_cvTune}
\title{Check for a \code{cvTune} object-}
\usage{
is_cvTune(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests if an object belong to \code{\link{cvTune}} class.
}
