% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modData_oop.R
\name{is_modData}
\alias{is_modData}
\alias{is_modMData}
\title{Test for an \code{modData} and \code{modMData} class instances.}
\usage{
is_modData(x)

is_modMData(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests is an object is an instance of the \code{modData} or \code{modMData} class.
}
