% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modHT_oop.R
\name{is_modHT}
\alias{is_modHT}
\title{Test for an \code{modHT} class instance.}
\usage{
is_modHT(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests is an object is an instance of the \code{modHT} class.
}
