% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTrain_oop.R
\name{is_modTrain}
\alias{is_modTrain}
\title{Test for an \code{modTrain} class instance.}
\usage{
is_modTrain(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests is an object is an instance of the \code{modTrain} class.
}
