% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat_stats}
\alias{mat_stats}
\title{Distribution statistics for a list of numeric matrices.}
\usage{
mat_stats(x)
}
\arguments{
\item{x}{a list of numeric matrices.}
}
\value{
a data frame with the statistics.
}
\description{
An internal tool used by \code{summary()} methods, which computes
numbers of features, observations, missing data poins and distribution
statistics for a list of numeric matrices.
}
