% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{modData}
\alias{modData}
\title{\code{modData} class.}
\usage{
modData(x, ...)
}
\arguments{
\item{x}{a list with elements specified in \code{Details}.}

\item{...}{extra arguments, currently none.}
}
\value{
An instance of \code{modData} class with the following methods specified:
\code{\link{nobs.modData}}, \code{\link{components.modData}},
\code{\link{summary.modData}}, and \code{\link{plot.modData}}.
}
\description{
Creates an object of \code{modData} class, which can be used for construction of
regularized regression models, e.g. with the \code{\link{pre_process}}
function.
}
\details{
Contains the following elements:
\itemize{
\item \code{train}: a batch-adjusted numeric training matrix of selected
explanatory factors, features in rows and observations in columns.
\item \code{test}: a batch-adjusted numeric test matrix of selected
explanatory factors, features in rows and observations in columns.
\item \code{stats}: distribution statistics of all features prior to the batch
adjustment: mean, median, Gini index,
and an indicator if the feature was selected as a modeling variable.
\item \code{mean_limits}: cutoffs of mean feature values used for selection of the
modeling variables.
\item \code{median_limits}: cutoffs of median feature values used for selection of
the modeling variables.
\item \code{gini_limits}: cutoffs of Gini coefficient used for selection of the
modeling variables.
}
}
