% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{modTrain}
\alias{modTrain}
\title{\code{modTrain} class.}
\usage{
modTrain(x, ...)
}
\arguments{
\item{x}{a list with the elements specified in \code{Details}.}

\item{...}{extra arguments, currently none.}
}
\value{
an instance of the \code{modTrain} class.
}
\description{
Creates an object of the \code{modTrain} class, which stores tuned
\code{glmnet} models and their basic performance stats.
It can be obtained by calling the
\code{\link{train.default}} or \code{\link{train.modData}} functions.
}
\details{
Contains the following elements:
\itemize{
\item \code{models}: a list of \code{\link[glmnet]{cv.glmnet}} models used for
tuning of \eqn{\lambda}, named after the columns of the \code{y} matrix.
\item \code{stats}: the optimal \eqn{lambda} parameter values, cross-validated
selection measures, and statistics of model performance in the
training data set.
\item \code{globals}: globals used for construction of the GLMNET models such as
\code{alpha} or \code{family}.
\item \code{errors}: a list of error messages for responses for which tuning/training
failed.
}

Note the \code{\link{nobs.modTrain}}, \code{\link{components.modTrain}},
\code{\link{summary.modTrain}}, \code{\link{plot.modTrain}},
\code{\link{coef.modTrain}}, and \code{\link{predict.modTrain}} methods
defined for the class.
}
