% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvTune_oop.R
\name{nobs.cvTune}
\alias{nobs.cvTune}
\title{Number of observations, features and non-zero features in \code{cvTune} objects.}
\usage{
\method{nobs}{cvTune}(object, ...)
}
\arguments{
\item{object}{an instance of the \code{\link{cvTune}} class.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with number of observations,
}
\description{
Retrieves the number of observations, all features, and features with
non-zero coefficients of the best performing \code{GLMNET} model.
}
