% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modData_oop.R
\name{nobs.modData}
\alias{nobs.modData}
\title{Numbers of variables and observations.}
\usage{
\method{nobs}{modData}(object, ...)
}
\arguments{
\item{object}{an instance of the \code{modData} class.}

\item{...}{extra arguments, currently none.}
}
\value{
a numeric vector.
}
\description{
Numbers of observations and variables in a \code{modData} object.
}
