% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTrain_oop.R
\name{nobs.modTrain}
\alias{nobs.modTrain}
\title{Numbers of observations, explanatory variables and models.}
\usage{
\method{nobs}{modTrain}(object, ...)
}
\arguments{
\item{object}{an instance of the \code{modTrain} class.}

\item{...}{extra arguments, currently none.}
}
\value{
a data frame with the response name,
numbers of complete observations, numbers of modeling features, and
numbers of non-zero coeffcients of the regularized models.
}
\description{
Computes numbers of complete observations, explanatory variables
and models in a \code{modTrain} object.
}
