% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paad_cptac}
\alias{paad_cptac}
\title{Bulk cancer RNA sequencing, PAAD CPTAC study.}
\format{
a numeric matrix with RNA sequencing gene expression levels normalized by
RSEM and log-transformed. The matrix contains 23642 rows and 140 columns.
Genes (symbols) are placed in rows, cancer sample identifiers are in columns.
}
\source{
https://www.cbioportal.org/study/summary?id=paad_cptac_2021
}
\usage{
data(paad_cptac)
}
\description{
Bulk cancer expression data (RSEM, log transformed) for 140 pancreatic
ductal cancer samples from the PAAD CPTAC study.
}
\references{
Cao L, Huang C, Cui Zhou D, Hu Y, Lih TM, Savage SR, Krug K, Clark DJ,
Schnaubelt M, Chen L, et al. Proteogenomic characterization of pancreatic
ductal adenocarcinoma. Cell (2021) 184:5031-5052.e26.
doi:10.1016/J.CELL.2021.08.023
}
