% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paad_tcga}
\alias{paad_tcga}
\title{Bulk cancer RNA sequencing, TCGA PAAD study.}
\format{
a numeric matrix with RNA sequencing gene expression levels normalized by
RSEM and log-transformed. The matrix contains 20511 rows and 177 columns.
Genes (symbols) are placed in rows, cancer sample identifiers are in columns.
}
\source{
https://www.cbioportal.org/study/summary?id=paad_tcga_pan_can_atlas_2018
}
\usage{
data(paad_tcga)
}
\description{
Bulk cancer expression data (RSEM, log transformed) for 177 pancreatic
ductal cancer samples from the pan-cancer TCGA PAAD cohort.
}
\references{
Liu J, Lichtenberg T, Hoadley KA, Poisson LM, Lazar AJ, Cherniack AD,
Kovatich AJ, Benz CC, Levine DA, Lee A V., et al.
An Integrated TCGA Pan-Cancer Clinical Data Resource to Drive High-Quality
Survival Outcome Analytics.
Cell (2018) 173:400-416.e11. doi:10.1016/J.CELL.2018.02.052
}
