% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvTune_oop.R
\name{plot.cvTune}
\alias{plot.cvTune}
\title{Diagnostic plots for tuninig of \code{GLMNET} models.}
\usage{
\method{plot}{cvTune}(
  x,
  palette = c("steelblue", "orangered3"),
  point_alpha = 0.75,
  point_size = 2,
  ...
)
}
\arguments{
\item{x}{an instance of \code{\link{cvTune}} class.}

\item{palette}{a two element vector of names of colors for, respectively,
all data points and the best tune.}

\item{point_alpha}{data point alpha.}

\item{point_size}{size of the data points.}

\item{...}{extra arguments, currently none.}
}
\value{
a \code{ggplot} graphics, which can be customized by the user.
}
\description{
Plots of the cost function values and \code{lambda} in repeats of the
cross-validated tuning of \code{GLMNET} models.
}
