% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modData_oop.R
\name{plot.modData}
\alias{plot.modData}
\title{Diagnostic plots for a \code{modData} object.}
\usage{
\method{plot}{modData}(
  x,
  palette = c(train = "indianred", test = "steelblue"),
  alpha = 0.5,
  stat_color = "orangered2",
  show_cutoffs = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{modData} object.}

\item{palette}{a named vector with colors for the training and test data sets.}

\item{alpha}{alpha of the violins.}

\item{stat_color}{color of the points and whiskers representing
medians and interquartile ranges.}

\item{show_cutoffs}{logical, if \code{TRUE}, cutoffs of mean, median and Gini
index used to select modeling variables will be displayed in plots.
Valid only for objects with a single test matrix.}

\item{...}{extra arguments passed to \code{\link[ggplot2]{geom_violin}}.}
}
\value{
a list of \code{ggplot} objects.
}
\description{
Generates a series of diagnostic plots for a \code{modData} object.
}
\details{
The following plots are returned:
\itemize{
\item violin plots of feature medians, means and Gini indexes prior to
the adjustment with the user-specified cutoffs denoted as dashed lines.
\item violin plots of feature medians and means after the batch effect
adjustment.
}
}
