% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modHT_oop.R, R/modTrain_oop.R
\name{plot.modHT}
\alias{plot.modHT}
\alias{plot.modTrain}
\title{Diagnostic plots.}
\usage{
\method{plot}{modHT}(
  x,
  shape_alpha = 0.5,
  shape_color = "steelblue",
  point_alpha = 0.75,
  point_color = shape_color,
  point_size = 2,
  point_wjitter = 0,
  point_hjitter = 0,
  show_txt = TRUE,
  txt_size = 2.75,
  ...
)

\method{plot}{modTrain}(
  x,
  shape_alpha = 0.5,
  shape_color = "steelblue",
  point_alpha = 0.75,
  point_color = shape_color,
  point_size = 2,
  point_wjitter = 0,
  point_hjitter = 0,
  show_txt = TRUE,
  txt_size = 2.75,
  ...
)
}
\arguments{
\item{x}{a \code{modTrain} or \verb{modHT object} object.}

\item{shape_alpha}{alpha for violin plot.}

\item{shape_color}{color of the violin plot.}

\item{point_alpha}{data point alpha.}

\item{point_color}{color of the data points.}

\item{point_size}{size of the data points.}

\item{point_wjitter}{jittering width of the data points.}

\item{point_hjitter}{jittering  height of the data points.}

\item{show_txt}{logical, should response names be displayed in the plots?}

\item{txt_size}{size of the text.}

\item{...}{extra arguments, currently none.}
}
\value{
a list of \code{ggplot} plots, appearance can be easily customized
by the user.
}
\description{
Diagnostic plots of model performance measures in cross-validation and in
the train data set.
}
\details{
The following plots are generated:
\itemize{
\item violin plot of the cross-validated performance statistic used for
selection of the optimal \eqn{\lambda}, \eqn{\R^2}, correlation coefficients,
and classification metrics.
\item optional scatter plots of cross-validated \eqn{\R^2} or cross-validated
loss function and correlation coefficients (Pearson and Spearman, training
data, Gaussian and Poisson family models), cross-validated loss function and
classification metrics (accuracy and Cohen's \eqn{\kappa}, training data,
classification models).
}
}
