% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_stats}
\alias{plot_stats}
\title{Model diagnostic plots.}
\usage{
plot_stats(
  data,
  family,
  shape_alpha = 0.5,
  shape_color = "steelblue",
  point_alpha = 0.75,
  point_color = shape_color,
  point_size = 2,
  point_wjitter = 0,
  point_hjitter = 0,
  show_txt = TRUE,
  txt_size = 2.75,
  txt_color = "black"
)
}
\arguments{
\item{data}{a data frame with model performance metrics.}

\item{family}{name of the modeling family.}

\item{shape_alpha}{alpha for violin plot.}

\item{shape_color}{color of the violin plot.}

\item{point_alpha}{data point alpha.}

\item{point_color}{color of the data points.}

\item{point_size}{size of the data points.}

\item{point_wjitter}{jittering width of the data points.}

\item{point_hjitter}{jittering  height of the data points.}

\item{show_txt}{logical, should response names be displayed in the plots?}

\item{txt_size}{size of the text.}

\item{txt_color}{color of the text.}
}
\value{
a list of \code{ggplot} objects.
}
\description{
A function intended for internal use that takes a data frame with plotting
stats and a modeling family name and returns family-specific violin and
acatter plots.
}
