% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_}
\alias{pred_}
\title{Predictions for a list \code{cv.glmnet} models.}
\usage{
pred_(models, lambdas, newx, type, family, newoffset = NULL, ...)
}
\arguments{
\item{models}{a list of \code{cv.glmnet} models.}

\item{lambdas}{a numeric vector of lambdas to make predictions for.}

\item{newx}{a numeric matrix of explanatory factors.}

\item{type}{of prediction, see: \code{\link{predict.modTrain}}.}

\item{family}{modeling family.}

\item{newoffset}{modeling offset.}

\item{...}{extra arguments passed to \code{\link[glmnet]{predict.glmnet}}.}
}
\value{
As described for \code{\link{predict.modTrain}}.
}
\description{
Predicts responses, linear predictor scores, coefficients and similar,
for a list of \code{cv.glmnet} objects. Intended for the internal use.
}
