% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTrain_oop.R
\name{predict.modTrain}
\alias{predict.modTrain}
\alias{coef.modTrain}
\title{Predict outcomes or compute coefficients.}
\usage{
\method{predict}{modTrain}(
  object,
  newdata = NULL,
  type = c("link", "response", "class", "coefficients"),
  ...
)

\method{coef}{modTrain}(object, ...)
}
\arguments{
\item{object}{an \code{modTrain} object.}

\item{newdata}{a numeric matrix of explanatory factors with features in rows
and observations in columns. Alternatively, a \code{\link{modData}} object
created e.g. with the \code{\link{pre_process}} function.}

\item{type}{type of prediction, see: \code{Details}.}

\item{...}{extra arguments passed to \code{\link[glmnet]{predict.glmnet}}
or to \code{\link[glmnet]{coef.glmnet}}.}
}
\value{
specific for the prediction type, see: \code{Details}.
If a \code{\link{modData}} object with multiple test data
(see: \code{\link{multi_process}}) sets will be
provided as \code{newdata}, a list of matrices is returned.
}
\description{
Predicts outcome, logits, coefficients and similar for a \code{modTrain} object.
}
\details{
For Gaussian family models with
types \code{type = 'link'} or \code{type = 'response'}, a matrix with predicted values
on the same scale as response is returned with user-specified transformations
and, optionally, standardization.
For other families, \code{type = 'link'} returns
a matrix of linear predictor score, while \code{type = 'response'} returns a
matrix of probabilities (binomial: one matrix, and multinomial regression:
list of matrices, with each element corresponding to one response variable)
or a matrix of predicted means (Poisson).
\code{type = 'class'} applies only to classification models and produces a matrix
of classes (labels). For \code{type = 'coefficients'} a matrix of coefficients for
the optimal \eqn{\lambda} is returned.
is returned.
See: \code{\link[glmnet]{predict.glmnet}} for details.
The \code{coef()} method returns a matrix of model coefficients for the optimal
\eqn{\lambda}.
}
