% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvTune_oop.R
\name{print.cvTune}
\alias{print.cvTune}
\title{Appearance of \code{cvTune} class.}
\usage{
\method{print}{cvTune}(x, ...)
}
\arguments{
\item{x}{an instance of \code{cvTune} class.}

\item{...}{extra arguments passed to methods.}
}
\value{
nothing, called for its side effects.
}
\description{
\code{print()} display method for \code{\link{cvTune}} objects.
}
