% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modData_oop.R
\name{print.modData}
\alias{print.modData}
\title{Appearance of the \code{modData} class.}
\usage{
\method{print}{modData}(x, ...)
}
\arguments{
\item{x}{an instance of the \code{modData} class.}

\item{...}{extra arguments, currently none.}
}
\value{
none, called for its side effects.
}
\description{
\code{print()} method for the \code{modData} class.
}
