% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modHT_oop.R
\name{print.modHT}
\alias{print.modHT}
\title{Appearance of the \code{modHT} class.}
\usage{
\method{print}{modHT}(x, ...)
}
\arguments{
\item{x}{an instance of the \code{modHT} class.}

\item{...}{extra arguments, currently none.}
}
\value{
none, called for its side effects.
}
\description{
\code{print()} method for the \code{modHT} class.
}
