% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTrain_oop.R
\name{print.modTrain}
\alias{print.modTrain}
\title{Appearance of the \code{modTrain} class.}
\usage{
\method{print}{modTrain}(x, ...)
}
\arguments{
\item{x}{an instance of the \code{modTrain} class.}

\item{...}{extra arguments, currently none.}
}
\value{
none, called for its side effects.
}
\description{
\code{print()} method for the \code{modTrain} class.
}
