% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvTune_oop.R
\name{summary.cvTune}
\alias{summary.cvTune}
\title{Performance statistics: summary of \code{cvTune} objects.}
\usage{
\method{summary}{cvTune}(object, ...)
}
\arguments{
\item{object}{an instance of \code{\link{cvTune}}.}

\item{...}{extra arguments passed to methods; currently none.}
}
\value{
a sorted data frame with the globally optimal \code{lambda} value in the first
row and the following columns:
\itemize{
\item \code{rep}: identification of a repeat
\item \code{cvm}: cross-validated cost function value for out-of-fold predictions in
particular repeats
\item \code{cvup} and \code{cvlo}: upper bond of 95\% confidence interval of \code{cvm}
}
}
\description{
Retrieves locally optimal \code{lambda} values with cross-validated cost function
values from the \code{\link{cvTune}} model.
}
