% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modHT_oop.R
\name{summary.modHT}
\alias{summary.modHT}
\title{Object summary.}
\usage{
\method{summary}{modHT}(object, ...)
}
\arguments{
\item{object}{a \code{modHT} object.}

\item{...}{extra argments, currently none.}
}
\value{
a data frame with model performance statistic in cross-validation
(loss function, optionally \eqn{\R^2}) and in the training data set
(optional: correlation coefficients, classification statistics),
along with numbers of observations, modeling variables, and
model coefficients.
}
\description{
Performance statistics of the models used for predictions in a \code{modHT} object.
}
