% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modTrain_oop.R
\name{summary.modTrain}
\alias{summary.modTrain}
\title{Object summary.}
\usage{
\method{summary}{modTrain}(object, ...)
}
\arguments{
\item{object}{a \code{modTrain} object.}

\item{...}{extra argments, currently none.}
}
\value{
a data frame with model performance statistic in cross-validation
(loss function, optionally \eqn{\R^2}) and in the training data set
(optional: correlation coefficients, classification statistics),
along with numbers of observations, modeling variables,
and model coefficients.
}
\description{
Performance statistics of the models stored in a \code{modTrain} object.
}
