% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{tune_glmnet}
\alias{tune_glmnet}
\title{Tune \code{GLMNET} models by repeated cross-validation.}
\usage{
tune_glmnet(
  x,
  y,
  fold_ids,
  type.measure = c("default", "mse", "deviance", "class", "auc", "mae", "C"),
  ...
)
}
\arguments{
\item{x}{numeric matrix or a \code{\link[stats]{model.matrix}} object
with explanatory variables; observations in rows and variables in columns.}

\item{y}{response object (numeric vector, factor, a matrix or s survival
object); please consult \code{\link[glmnet]{glmnet}} for details.}

\item{fold_ids}{a list of numeric vectors with indexes (one per repeat),
which specify the assignment of observations to cross-validation folds.
See: \code{\link[glmnet]{cv.glmnet}} for details and
\code{\link[caret]{createFolds}} for a handy function returning the
fold-assortment vectors for each of the repeats.}

\item{type.measure}{cost function as specified for
\code{\link[glmnet]{cv.glmnet}}. For \code{type.measure = 'default'}, the model
deviance is minimized.}

\item{...}{other arguments passed to \code{\link[glmnet]{glmnet}} and
\code{\link[glmnet]{cv.glmnet}}.}
}
\value{
a \code{\link{cvTune}} cvlass object with the following components:
\itemize{
\item \code{stats}: model performance statistic, i.e. the cost function, for the
locally optimal \code{lambda} values in repeats of the cross-validation. Those
are the \code{lambda} values, one per repeat, which correspond to the optimum of
the cost function in out-of-fold predictions. In case there are two local
optima, the first one corresponding to a smaller \code{lambda} is returned.
\item \code{lambda}: the globally optimal value of \code{lambda}m, i.e. the optimum of
\code{lambdas} for particular repeats listed in \code{stats}. If multiple values of
cost function are available for the same value of lambda, then the cost
function is averaged per lambda value prior to finding the optimal lambda
value.
\item \code{model}: a \code{GLMNET} model fit by \code{\link[glmnet]{glmnet}} for the
globally optimal \code{lambda} value
}

Note the methods \code{\link{nobs.cvTune}}, \code{\link{summary.cvTune}},
\code{\link{plot.cvTune}}, and \code{\link{predict.cvTune}} defined
for the class.
}
\description{
Finds the globally optimal value of the shrinkage parameter \code{lambda} for
regularized linear \code{GLMNET} models by minimizing or maximizing a cost
function in repeated cross-validation.
}
\details{
In principle, there is only one tuning parameter in regularized linear models
build by modeling tools of the \code{GLMNET} package: the shrinkage parameter
\code{lambda}, whose optimal value can be estimates in a repeated cross-validation
setting. Technically, this task is reduced to running
\code{\link[glmnet]{cv.glmnet}} for the requested number of repeats.
The function will run in parallel if a parallel backend is specified via
\code{\link[future]{plan}}.
}
