% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{x_mats}
\alias{x_mats}
\title{Pre-process a list of numeric matrices.}
\usage{
x_mats(
  x,
  mean_quantile = NULL,
  median_quantile = NULL,
  gini_quantile = NULL,
  trans_fun = identity,
  ...
)
}
\arguments{
\item{x}{a list of numeric matrices. Modeling features are provided in rows,
observations are in columns.#'}

\item{mean_quantile}{a numeric that specifies the quantile cutoff
of mean value of the features to be selected. Default is \code{NULL}, which means
that no filtering by mean feature value will be applied.}

\item{median_quantile}{as above, a quantile of median used to select
the features. Defaults to \code{NULL}, which implicates that no median feature
filtering will be applied.}

\item{gini_quantile}{as above, a quantile of Gini coefficient used to select
the modeling features of sufficient variability. If \code{NULL} (default), no
filtering of modeling features by Gini coefficient will be applied.}

\item{trans_fun}{a function used to transform the output,
\code{identity} by default.}

\item{...}{extra arguments passed to \code{\link[sva]{ComBat}}.}
}
\value{
A list with two elements:
\itemize{
\item \code{x}: a list of numeric matrices adjusted for bacth effects and containing
only the selected modeling features.
\item \code{stats}: feature distributions statistics (median, mean, Gini index) of
all features shared by the input data.
\item \code{mean_limits}, \code{median_limits}, \code{gini_limits}: numeric vectors of cutoffs
of mean, median, and Gini index of explanatory features used for selection
of the modeling variables.
}
}
\description{
Pre-processing of a list of numeric matrices as described for
\code{\link{pre_process}}.
}
\details{
For internal use only!
}
