% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modData_oop.R
\name{is_modData}
\alias{is_modData}
\title{Test for an \code{modData} class instance.}
\usage{
is_modData(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Tests is an object is an instance of the \code{modData} class.
}
